use the `int` type for progress within the range [0-100]

+++ src/main/java/com/creactiviti/piper/core/error/Progressable.java
@@ -8,7 +8,7 @@ package com.creactiviti.piper.core.error;
  */
 public interface Progressable {
     /**
-     *  @return the current progress value, a number between 0. and 1., or <code>null</code> otherwise.
+     *  @return the current progress value, a number between 0 and 100.
      */
-    Float getProgress ();
+    int getProgress ();
 }
